refactor: narrow wildcard module surfaces
